Fitness_brand_case_study
A dataset is given from a fitness brand with information about the sales of their 3 treadmill models.
Descriptive as well as analytical analysis is done on the dataset to discern target audience characteristics for each treadmill variant, optimizing recommendations for new customers. Employing descriptive analytics, we develop comprehensive customer profiles using tables and charts. Additionally, two-way contingency tables are constructed for each treadmill, facilitating the computation of conditional and marginal probabilities. The insights derived from this analysis are pivotal for informed decision-making and have a substantial impact on refining business strategies.